requirements.txt
requests==2.27.1
joblib==1.1.0
pandas==1.4.2
pyyaml==6.0
scikit-learn==1.0.2
xgboost==1.6.1